+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Change log
 Next version
 ------------
 
+0.20
+----
+
 * Added Django 4.1 to the CI matrix.
 * Extended the ``email`` action with the optional ability to add the author of
   the sibmission to the Cc: of the sent email.